move suggeestions to container
