+++ src/container/Login/login.scss
@@ -8,7 +8,8 @@
   left: 50%;
   transform: translate(-50%, -50%);
   border-radius: 6px;
-  box-shadow: 0 0 4px #c4c4c4;
+  // box-shadow: 0 0 4px #c4c4c4;
+  border: 1px solid;
 
   .buttons-container {
     .login-header {
